can put and deleter review

+++ controllers/reviewController.js
@@ -61,12 +61,54 @@ export async function getMyReview(req, res) {
     }
 }
 
+export async function updateReview(req, res) {
+    console.log("PUT /review is requested");
+    try {
+        if (req.params.username == null  ) {
+            console.log("Fail in null");
+            return res.json({ messageUpdateReview: "fail" });
+        }
+        const existsResult = await database.query({
+            text: `SELECT * FROM reviews WHERE "reviewId" = $1`,
+            values: [req.body.reviewId],
+        });
+
+        if(!existsResult.rows[0]) {
+            console.log("fail in exists");
+            return res.json({ messageUpdateReview: "fail" });
+        }
+        const currentDate = new Date();
+        const result = await database.query({
+            text: `UPDATE reviews SET "score" = $1, "comment" = $2 , "reviewUpdateDate" = $3 WHERE "reviewId" = $4`,
+            values: [req.body.score, req.body.comment, currentDate ,req.body.reviewId],
+        });
+        console.log("success");
+        return res.json({ messageUpdateReview: "success" });
+    } catch (err) {
+        console.log("Error in catch", err);
+        return res.json({ messageUpdateReview: "fail" });
+    }
+}
+
 export async function deleteReview(req, res) {
     console.log("DELETE /review is requested");
     try {
+        if (req.params.username == null ) {
+            console.log("Fail in null");
+            return res.json({ messageDeleteReview: "fail" });
+        }
+
+        const existsResult = await database.query({
+            text: `SELECT * FROM reviews WHERE "reviewId" = $1`,
+            values: [req.body.reviewId],
+        });
+        if(!existsResult.rows[0]) {
+            console.log("fail in exists");
+            return res.json({ messageDeleteReview: "fail" });
+        }
         const result = await database.query({
             text: `DELETE FROM reviews WHERE "reviewId" = $1`,
-            values: [req.params.reviewId],
+            values: [req.body.reviewId],
         });
         console.log("success");
         return res.json({ messageDeleteReview: "success" });
